+++ NEWS.md
@@ -27,8 +27,7 @@ colors are very distinct from each other).
 `geom_point`. This can be used to avoid overplotting. (#67)
 * `overlay_function` convenience function. Example usage: add a Gaussian (or any
 distribution) density curve to a plot made with `mcmc_hist`.
-* `mcmc_recover_scatter`, which is similar to `mcmc_recover_intervals` but plots
-estimated vs "true" values as a scatterplot. (#81)
+* `mcmc_recover_scatter` and `mcmc_recover_hist`, which are similar to `mcmc_recover_intervals` and compare estimates to "true" values used to simulate data. (#81, #83)
 * New PPC category __Discrete__ with functions: 
     - `ppc_rootogram` for use with models for count data. Thanks to @paul-buerkner. (#28)
     - `ppc_bars`, `ppc_bars_grouped` for use with models for ordinal, categorical